+++ sdk/docs/ref/testing/functions/setupMockServer.md
@@ -1,4 +1,4 @@
-[**@saflib/vue**](../../../../index.md)
+[**@saflib/sdk**](../../index.md)
 
 ***
