+++ sdk/docs/ref/testing/index.md
@@ -8,4 +8,6 @@
 
 | Function | Description |
 | ------ | ------ |
+| [setupMockServer](functions/setupMockServer.md) | Simple wrapper around `msw`'s `setupServer` function. |
 | [typedCreateHandler](functions/typedCreateHandler.md) | Use to create a typed helper function for creating typesafe mock API handlers. |
+| [withVueQuery](functions/withVueQuery.md) | Helper function to test Vue Query composables in isolation. |
