+++ vue/docs/ref/@saflib/vue/testing/index.md
@@ -18,6 +18,5 @@ Testing utilities for Vue.
 | ------ | ------ |
 | [getElementByString](functions/getElementByString.md) | This should always be used to find elements in tests. |
 | [mountWithPlugins](functions/mountWithPlugins.md) | Mount a Vue component with plugins. Handles plugins like vuetify, router, and i18n. Uses `mount` under the hood. |
-| [setupMockServer](functions/setupMockServer.md) | Simple wrapper around `msw`'s `setupServer` function. |
 | [stubGlobals](functions/stubGlobals.md) | Call during test setup to stub browser globals like ResizeObserver, matchMedia, location, and visualViewport. |
 | [withVueQuery](functions/withVueQuery.md) | Helper function to test Vue Query composables in isolation. |
