+++ vue/testing/index.ts
@@ -4,6 +4,6 @@
  */
 
 export * from "./mount.ts";
-export * from "./requests.ts";
+export * from "../../sdk/testing/with-vue-query.ts";
 export * from "./string-utils.ts";
 export * from "./globals.ts";
